Rewrite VPG agent and fix training bug
OK! So, I've been able to understand a lot of the most basic bugs I have
been having with Snake over the past week!
During reward_to_go calculation we had this code:

But that last line is wrong! It might seem like we need to reverse the
list, because we are iterating in reverse order up above, but in fact we
are correctly walking the reward series backwards, we shouldn't reverse
it again at the end!
This means that we went from something like this:

```
[
  ([...], 2, -1.),
  ([...], 1, 1.),
]
```

To this:

```
[
  ([...], 2, 1.),
  ([...], 1, -1.),
]
```

Notice: We have switched the rewards and the actions that earned them!
When we are running contrived data through the network this will cause
really wacky things to happen (because the contrived actions and rewards
will actually result in opposite training), whereas during training on
real data where rewards are sparse and experience histories are long
this will result in basically training on noise (which is what we were
seeing last night).