Changes ---
If all probabilities are 0, then bump everything up to 1.0 because
calling multinomial raises an error if everything totals to zero but not
if everything is the same value (which is what we care about).

Side note: I wonder if the surprise I was seeing logged a lot (15.924
or something like that) is the surprise when the probs are all 0?

If the surprise is too close to zero (e.g. 1.12e-7) then bump it up to
.05 so that at least *some* loss gets calculated.

This has the effect of never allowing behaviors to become completely
burned in! Another option would be to do this only for negative reward
scenarios, but I think it's fine to just do it all around.